EEG Python/requirements.txt
Pillow
pandas
numpy
networkx
matplotlib
scipy
glob2
XlsxWriter
